+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to
 - Support for nodeinfo 2.1 and 2.2.
 - ActorStorage method for checking active users.
 - nodeinfo properties for monthly active and half-yearly active users.
+- SafeFetcher service class for all outbound HTTP requests
 
 ### Changed