+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to
 
 - Name of package in nodeinfo from `activitypub-bot` to `activitypub-dot-bot` to
   avoid a naming conflict with another package.
-- Use URLFormatter in Transformer to short-circuit Webfinger self-request
+- Use URLFormatter in Transformer and BotContext to short-circuit Webfinger self-request
 - Use SafeFetcher in ActivityPubClient
 - Use SafeFetcher in Transformer
 - Use SafeFetcher in BotContext